xQTLdownload_eqtlSig

download significant eQTL associations of a tissue or across all tissues.

Usage

xQTLdownload_eqtlSig(
  variantName = "",
  genes = "",
  variantType = "auto",
  geneType = "auto",
  tissueSiteDetail = "",
  datasetId = "gtex_v8"
)

Arguments

variantName

A character string. like dbsnp ID or variant id in GTEx.

genes

A gene symbol or a gencode id (versioned).

variantType

A character string. "auto", "snpId" or "variantId". Default: "auto".

geneType

A character string. "auto","geneSymbol" or "gencodeId". Default: "auto".

tissueSiteDetail

A character string. tissue detail can be listed using \"tissueSiteDetailGTExv8\" or \"tissueSiteDetailGTExv7\"

datasetId

A character string. "gtex_v8" or "gtex_v7". Default: "gtex_v8".

Value

A data.table

Examples

# \donttest{
 # Download eQTL info for a variant:
 xQTLdownload_eqtlSig("rs201327123")
#> == Check the variant name:
#> Querying variant... | 2022-6月-01 16:34:55 
#> == Done.
#> == Start downloading significant eQTL associations:
#> == Done.
#> =================================
#> Totally 71 associatons were found for variant [rs201327123] in 32 tissues in gtex_v8.
#>              variantId       snpId         gencodeId    geneSymbol
#>  1: chr1_14677_G_A_b38 rs201327123 ENSG00000241860.6 RP11-34P13.13
#>  2: chr1_14677_G_A_b38 rs201327123 ENSG00000228463.9    AP006222.2
#>  3: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#>  4: chr1_14677_G_A_b38 rs201327123 ENSG00000241860.6 RP11-34P13.13
#>  5: chr1_14677_G_A_b38 rs201327123 ENSG00000228463.9    AP006222.2
#>  6: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#>  7: chr1_14677_G_A_b38 rs201327123 ENSG00000241860.6 RP11-34P13.13
#>  8: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#>  9: chr1_14677_G_A_b38 rs201327123 ENSG00000241860.6 RP11-34P13.13
#> 10: chr1_14677_G_A_b38 rs201327123 ENSG00000228463.9    AP006222.2
#> 11: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#> 12: chr1_14677_G_A_b38 rs201327123 ENSG00000241860.6 RP11-34P13.13
#> 13: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#> 14: chr1_14677_G_A_b38 rs201327123 ENSG00000241860.6 RP11-34P13.13
#> 15: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#> 16: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#> 17: chr1_14677_G_A_b38 rs201327123 ENSG00000241860.6 RP11-34P13.13
#> 18: chr1_14677_G_A_b38 rs201327123 ENSG00000228463.9    AP006222.2
#> 19: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#> 20: chr1_14677_G_A_b38 rs201327123 ENSG00000241860.6 RP11-34P13.13
#> 21: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#> 22: chr1_14677_G_A_b38 rs201327123 ENSG00000228463.9    AP006222.2
#> 23: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#> 24: chr1_14677_G_A_b38 rs201327123 ENSG00000241860.6 RP11-34P13.13
#> 25: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#> 26: chr1_14677_G_A_b38 rs201327123 ENSG00000241860.6 RP11-34P13.13
#> 27: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#> 28: chr1_14677_G_A_b38 rs201327123 ENSG00000241860.6 RP11-34P13.13
#> 29: chr1_14677_G_A_b38 rs201327123 ENSG00000228463.9    AP006222.2
#> 30: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#> 31: chr1_14677_G_A_b38 rs201327123 ENSG00000241860.6 RP11-34P13.13
#> 32: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#> 33: chr1_14677_G_A_b38 rs201327123 ENSG00000241860.6 RP11-34P13.13
#> 34: chr1_14677_G_A_b38 rs201327123 ENSG00000228463.9    AP006222.2
#> 35: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#> 36: chr1_14677_G_A_b38 rs201327123 ENSG00000241860.6 RP11-34P13.13
#> 37: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#> 38: chr1_14677_G_A_b38 rs201327123 ENSG00000241860.6 RP11-34P13.13
#> 39: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#> 40: chr1_14677_G_A_b38 rs201327123 ENSG00000241860.6 RP11-34P13.13
#> 41: chr1_14677_G_A_b38 rs201327123 ENSG00000228463.9    AP006222.2
#> 42: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#> 43: chr1_14677_G_A_b38 rs201327123 ENSG00000241860.6 RP11-34P13.13
#> 44: chr1_14677_G_A_b38 rs201327123 ENSG00000228463.9    AP006222.2
#> 45: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#> 46: chr1_14677_G_A_b38 rs201327123 ENSG00000228463.9    AP006222.2
#> 47: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#> 48: chr1_14677_G_A_b38 rs201327123 ENSG00000228463.9    AP006222.2
#> 49: chr1_14677_G_A_b38 rs201327123 ENSG00000241860.6 RP11-34P13.13
#> 50: chr1_14677_G_A_b38 rs201327123 ENSG00000228463.9    AP006222.2
#> 51: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#> 52: chr1_14677_G_A_b38 rs201327123 ENSG00000228463.9    AP006222.2
#> 53: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#> 54: chr1_14677_G_A_b38 rs201327123 ENSG00000241860.6 RP11-34P13.13
#> 55: chr1_14677_G_A_b38 rs201327123 ENSG00000228463.9    AP006222.2
#> 56: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#> 57: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#> 58: chr1_14677_G_A_b38 rs201327123 ENSG00000241860.6 RP11-34P13.13
#> 59: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#> 60: chr1_14677_G_A_b38 rs201327123 ENSG00000241860.6 RP11-34P13.13
#> 61: chr1_14677_G_A_b38 rs201327123 ENSG00000241860.6 RP11-34P13.13
#> 62: chr1_14677_G_A_b38 rs201327123 ENSG00000228463.9    AP006222.2
#> 63: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#> 64: chr1_14677_G_A_b38 rs201327123 ENSG00000241860.6 RP11-34P13.13
#> 65: chr1_14677_G_A_b38 rs201327123 ENSG00000228463.9    AP006222.2
#> 66: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#> 67: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#> 68: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#> 69: chr1_14677_G_A_b38 rs201327123 ENSG00000241860.6 RP11-34P13.13
#> 70: chr1_14677_G_A_b38 rs201327123 ENSG00000228463.9    AP006222.2
#> 71: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#>              variantId       snpId         gencodeId    geneSymbol
#>                              tissueSiteDetail      pValue       nes datasetId
#>  1:                    Adipose - Subcutaneous 7.91948e-08  0.700658   gtex_v8
#>  2:                    Adipose - Subcutaneous 8.21597e-10 -0.705895   gtex_v8
#>  3:                    Adipose - Subcutaneous 2.12632e-15  0.742337   gtex_v8
#>  4:              Adipose - Visceral (Omentum) 3.45758e-10  0.955911   gtex_v8
#>  5:              Adipose - Visceral (Omentum) 5.71926e-13 -0.901493   gtex_v8
#>  6:              Adipose - Visceral (Omentum) 2.70834e-14  0.889035   gtex_v8
#>  7:                            Artery - Aorta 6.62770e-07  0.844491   gtex_v8
#>  8:                            Artery - Aorta 6.26212e-07  0.615087   gtex_v8
#>  9:                           Artery - Tibial 1.94366e-11  0.924220   gtex_v8
#> 10:                           Artery - Tibial 3.96245e-06 -0.625202   gtex_v8
#> 11:                           Artery - Tibial 1.41430e-08  0.445997   gtex_v8
#> 12:           Brain - Caudate (basal ganglia) 5.56630e-05  0.855631   gtex_v8
#> 13:           Brain - Caudate (basal ganglia) 3.88785e-06  0.723820   gtex_v8
#> 14:             Brain - Cerebellar Hemisphere 5.66701e-05  1.062420   gtex_v8
#> 15:             Brain - Cerebellar Hemisphere 6.41620e-06  1.005240   gtex_v8
#> 16:                        Brain - Cerebellum 9.20661e-07  1.043500   gtex_v8
#> 17:                            Brain - Cortex 3.97305e-07  1.189090   gtex_v8
#> 18:                            Brain - Cortex 4.70573e-05 -0.911460   gtex_v8
#> 19:                            Brain - Cortex 1.14069e-08  1.143840   gtex_v8
#> 20:              Brain - Frontal Cortex (BA9) 1.93296e-06  1.076710   gtex_v8
#> 21:              Brain - Frontal Cortex (BA9) 3.69825e-07  1.016420   gtex_v8
#> 22:                      Brain - Hypothalamus 7.89694e-06 -1.058490   gtex_v8
#> 23:                      Brain - Hypothalamus 9.19176e-07  0.972513   gtex_v8
#> 24: Brain - Nucleus accumbens (basal ganglia) 6.43853e-08  1.124170   gtex_v8
#> 25: Brain - Nucleus accumbens (basal ganglia) 2.52087e-07  0.879910   gtex_v8
#> 26:                   Breast - Mammary Tissue 3.10205e-06  0.782287   gtex_v8
#> 27:                   Breast - Mammary Tissue 2.59853e-05  0.519586   gtex_v8
#> 28:              Cells - Cultured fibroblasts 3.16761e-09  0.870255   gtex_v8
#> 29:              Cells - Cultured fibroblasts 5.56059e-07 -0.701840   gtex_v8
#> 30:              Cells - Cultured fibroblasts 7.52028e-12  0.634564   gtex_v8
#> 31:                           Colon - Sigmoid 7.67178e-07  0.944220   gtex_v8
#> 32:                           Colon - Sigmoid 1.23965e-05  0.631243   gtex_v8
#> 33:                        Colon - Transverse 6.07252e-07  1.030600   gtex_v8
#> 34:                        Colon - Transverse 2.83738e-05 -0.807644   gtex_v8
#> 35:                        Colon - Transverse 2.34274e-07  0.740212   gtex_v8
#> 36:     Esophagus - Gastroesophageal Junction 2.71848e-07  1.001690   gtex_v8
#> 37:     Esophagus - Gastroesophageal Junction 9.76159e-07  0.634624   gtex_v8
#> 38:                        Esophagus - Mucosa 2.05779e-04  0.471568   gtex_v8
#> 39:                        Esophagus - Mucosa 2.48068e-09  0.655426   gtex_v8
#> 40:                    Esophagus - Muscularis 3.54545e-10  0.984300   gtex_v8
#> 41:                    Esophagus - Muscularis 6.11870e-07 -0.721811   gtex_v8
#> 42:                    Esophagus - Muscularis 2.50817e-05  0.408854   gtex_v8
#> 43:                  Heart - Atrial Appendage 2.12053e-08  1.000880   gtex_v8
#> 44:                  Heart - Atrial Appendage 4.72450e-10 -1.089990   gtex_v8
#> 45:                  Heart - Atrial Appendage 1.34356e-10  0.905711   gtex_v8
#> 46:                    Heart - Left Ventricle 9.72458e-12 -1.034280   gtex_v8
#> 47:                    Heart - Left Ventricle 3.44863e-09  0.759825   gtex_v8
#> 48:                                     Liver 1.22632e-06 -0.992022   gtex_v8
#> 49:                                      Lung 1.32367e-07  0.747842   gtex_v8
#> 50:                                      Lung 4.68334e-05 -0.587194   gtex_v8
#> 51:                                      Lung 4.57721e-11  0.701631   gtex_v8
#> 52:                         Muscle - Skeletal 9.91823e-09 -0.620506   gtex_v8
#> 53:                         Muscle - Skeletal 1.45337e-14  0.666372   gtex_v8
#> 54:                            Nerve - Tibial 2.97619e-15  1.050980   gtex_v8
#> 55:                            Nerve - Tibial 2.01162e-11 -0.861511   gtex_v8
#> 56:                            Nerve - Tibial 1.25308e-13  0.653294   gtex_v8
#> 57:                                  Pancreas 1.48904e-05  0.771603   gtex_v8
#> 58:                                 Pituitary 5.39767e-05  0.882921   gtex_v8
#> 59:                                 Pituitary 6.13186e-08  0.945009   gtex_v8
#> 60:                                  Prostate 1.52886e-05  1.005930   gtex_v8
#> 61:       Skin - Not Sun Exposed (Suprapubic) 6.03296e-11  0.911987   gtex_v8
#> 62:       Skin - Not Sun Exposed (Suprapubic) 7.22930e-08 -0.770752   gtex_v8
#> 63:       Skin - Not Sun Exposed (Suprapubic) 1.72014e-14  0.699855   gtex_v8
#> 64:            Skin - Sun Exposed (Lower leg) 8.00361e-11  0.857182   gtex_v8
#> 65:            Skin - Sun Exposed (Lower leg) 3.30591e-09 -0.746043   gtex_v8
#> 66:            Skin - Sun Exposed (Lower leg) 1.23187e-15  0.688242   gtex_v8
#> 67:                                    Spleen 3.15980e-06  0.781740   gtex_v8
#> 68:                                   Stomach 1.57231e-05  0.625518   gtex_v8
#> 69:                                   Thyroid 7.31007e-08  0.767395   gtex_v8
#> 70:                                   Thyroid 6.40558e-08 -0.795998   gtex_v8
#> 71:                                   Thyroid 6.91838e-11  0.605093   gtex_v8
#>                              tissueSiteDetail      pValue       nes datasetId
 xQTLdownload_eqtlSig("chr1_14677_G_A_b38")
#> == Check the variant name:
#> Querying variant... | 2022-6月-01 16:34:56 
#> == Done.
#> == Start downloading significant eQTL associations:
#> == Done.
#> =================================
#> Totally 71 associatons were found for variant [chr1_14677_G_A_b38] in 32 tissues in gtex_v8.
#>              variantId       snpId         gencodeId    geneSymbol
#>  1: chr1_14677_G_A_b38 rs201327123 ENSG00000241860.6 RP11-34P13.13
#>  2: chr1_14677_G_A_b38 rs201327123 ENSG00000228463.9    AP006222.2
#>  3: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#>  4: chr1_14677_G_A_b38 rs201327123 ENSG00000241860.6 RP11-34P13.13
#>  5: chr1_14677_G_A_b38 rs201327123 ENSG00000228463.9    AP006222.2
#>  6: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#>  7: chr1_14677_G_A_b38 rs201327123 ENSG00000241860.6 RP11-34P13.13
#>  8: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#>  9: chr1_14677_G_A_b38 rs201327123 ENSG00000241860.6 RP11-34P13.13
#> 10: chr1_14677_G_A_b38 rs201327123 ENSG00000228463.9    AP006222.2
#> 11: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#> 12: chr1_14677_G_A_b38 rs201327123 ENSG00000241860.6 RP11-34P13.13
#> 13: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#> 14: chr1_14677_G_A_b38 rs201327123 ENSG00000241860.6 RP11-34P13.13
#> 15: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#> 16: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#> 17: chr1_14677_G_A_b38 rs201327123 ENSG00000241860.6 RP11-34P13.13
#> 18: chr1_14677_G_A_b38 rs201327123 ENSG00000228463.9    AP006222.2
#> 19: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#> 20: chr1_14677_G_A_b38 rs201327123 ENSG00000241860.6 RP11-34P13.13
#> 21: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#> 22: chr1_14677_G_A_b38 rs201327123 ENSG00000228463.9    AP006222.2
#> 23: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#> 24: chr1_14677_G_A_b38 rs201327123 ENSG00000241860.6 RP11-34P13.13
#> 25: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#> 26: chr1_14677_G_A_b38 rs201327123 ENSG00000241860.6 RP11-34P13.13
#> 27: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#> 28: chr1_14677_G_A_b38 rs201327123 ENSG00000241860.6 RP11-34P13.13
#> 29: chr1_14677_G_A_b38 rs201327123 ENSG00000228463.9    AP006222.2
#> 30: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#> 31: chr1_14677_G_A_b38 rs201327123 ENSG00000241860.6 RP11-34P13.13
#> 32: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#> 33: chr1_14677_G_A_b38 rs201327123 ENSG00000241860.6 RP11-34P13.13
#> 34: chr1_14677_G_A_b38 rs201327123 ENSG00000228463.9    AP006222.2
#> 35: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#> 36: chr1_14677_G_A_b38 rs201327123 ENSG00000241860.6 RP11-34P13.13
#> 37: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#> 38: chr1_14677_G_A_b38 rs201327123 ENSG00000241860.6 RP11-34P13.13
#> 39: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#> 40: chr1_14677_G_A_b38 rs201327123 ENSG00000241860.6 RP11-34P13.13
#> 41: chr1_14677_G_A_b38 rs201327123 ENSG00000228463.9    AP006222.2
#> 42: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#> 43: chr1_14677_G_A_b38 rs201327123 ENSG00000241860.6 RP11-34P13.13
#> 44: chr1_14677_G_A_b38 rs201327123 ENSG00000228463.9    AP006222.2
#> 45: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#> 46: chr1_14677_G_A_b38 rs201327123 ENSG00000228463.9    AP006222.2
#> 47: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#> 48: chr1_14677_G_A_b38 rs201327123 ENSG00000228463.9    AP006222.2
#> 49: chr1_14677_G_A_b38 rs201327123 ENSG00000241860.6 RP11-34P13.13
#> 50: chr1_14677_G_A_b38 rs201327123 ENSG00000228463.9    AP006222.2
#> 51: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#> 52: chr1_14677_G_A_b38 rs201327123 ENSG00000228463.9    AP006222.2
#> 53: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#> 54: chr1_14677_G_A_b38 rs201327123 ENSG00000241860.6 RP11-34P13.13
#> 55: chr1_14677_G_A_b38 rs201327123 ENSG00000228463.9    AP006222.2
#> 56: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#> 57: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#> 58: chr1_14677_G_A_b38 rs201327123 ENSG00000241860.6 RP11-34P13.13
#> 59: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#> 60: chr1_14677_G_A_b38 rs201327123 ENSG00000241860.6 RP11-34P13.13
#> 61: chr1_14677_G_A_b38 rs201327123 ENSG00000241860.6 RP11-34P13.13
#> 62: chr1_14677_G_A_b38 rs201327123 ENSG00000228463.9    AP006222.2
#> 63: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#> 64: chr1_14677_G_A_b38 rs201327123 ENSG00000241860.6 RP11-34P13.13
#> 65: chr1_14677_G_A_b38 rs201327123 ENSG00000228463.9    AP006222.2
#> 66: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#> 67: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#> 68: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#> 69: chr1_14677_G_A_b38 rs201327123 ENSG00000241860.6 RP11-34P13.13
#> 70: chr1_14677_G_A_b38 rs201327123 ENSG00000228463.9    AP006222.2
#> 71: chr1_14677_G_A_b38 rs201327123 ENSG00000228327.3 RP11-206L10.2
#>              variantId       snpId         gencodeId    geneSymbol
#>                              tissueSiteDetail      pValue       nes datasetId
#>  1:                    Adipose - Subcutaneous 7.91948e-08  0.700658   gtex_v8
#>  2:                    Adipose - Subcutaneous 8.21597e-10 -0.705895   gtex_v8
#>  3:                    Adipose - Subcutaneous 2.12632e-15  0.742337   gtex_v8
#>  4:              Adipose - Visceral (Omentum) 3.45758e-10  0.955911   gtex_v8
#>  5:              Adipose - Visceral (Omentum) 5.71926e-13 -0.901493   gtex_v8
#>  6:              Adipose - Visceral (Omentum) 2.70834e-14  0.889035   gtex_v8
#>  7:                            Artery - Aorta 6.62770e-07  0.844491   gtex_v8
#>  8:                            Artery - Aorta 6.26212e-07  0.615087   gtex_v8
#>  9:                           Artery - Tibial 1.94366e-11  0.924220   gtex_v8
#> 10:                           Artery - Tibial 3.96245e-06 -0.625202   gtex_v8
#> 11:                           Artery - Tibial 1.41430e-08  0.445997   gtex_v8
#> 12:           Brain - Caudate (basal ganglia) 5.56630e-05  0.855631   gtex_v8
#> 13:           Brain - Caudate (basal ganglia) 3.88785e-06  0.723820   gtex_v8
#> 14:             Brain - Cerebellar Hemisphere 5.66701e-05  1.062420   gtex_v8
#> 15:             Brain - Cerebellar Hemisphere 6.41620e-06  1.005240   gtex_v8
#> 16:                        Brain - Cerebellum 9.20661e-07  1.043500   gtex_v8
#> 17:                            Brain - Cortex 3.97305e-07  1.189090   gtex_v8
#> 18:                            Brain - Cortex 4.70573e-05 -0.911460   gtex_v8
#> 19:                            Brain - Cortex 1.14069e-08  1.143840   gtex_v8
#> 20:              Brain - Frontal Cortex (BA9) 1.93296e-06  1.076710   gtex_v8
#> 21:              Brain - Frontal Cortex (BA9) 3.69825e-07  1.016420   gtex_v8
#> 22:                      Brain - Hypothalamus 7.89694e-06 -1.058490   gtex_v8
#> 23:                      Brain - Hypothalamus 9.19176e-07  0.972513   gtex_v8
#> 24: Brain - Nucleus accumbens (basal ganglia) 6.43853e-08  1.124170   gtex_v8
#> 25: Brain - Nucleus accumbens (basal ganglia) 2.52087e-07  0.879910   gtex_v8
#> 26:                   Breast - Mammary Tissue 3.10205e-06  0.782287   gtex_v8
#> 27:                   Breast - Mammary Tissue 2.59853e-05  0.519586   gtex_v8
#> 28:              Cells - Cultured fibroblasts 3.16761e-09  0.870255   gtex_v8
#> 29:              Cells - Cultured fibroblasts 5.56059e-07 -0.701840   gtex_v8
#> 30:              Cells - Cultured fibroblasts 7.52028e-12  0.634564   gtex_v8
#> 31:                           Colon - Sigmoid 7.67178e-07  0.944220   gtex_v8
#> 32:                           Colon - Sigmoid 1.23965e-05  0.631243   gtex_v8
#> 33:                        Colon - Transverse 6.07252e-07  1.030600   gtex_v8
#> 34:                        Colon - Transverse 2.83738e-05 -0.807644   gtex_v8
#> 35:                        Colon - Transverse 2.34274e-07  0.740212   gtex_v8
#> 36:     Esophagus - Gastroesophageal Junction 2.71848e-07  1.001690   gtex_v8
#> 37:     Esophagus - Gastroesophageal Junction 9.76159e-07  0.634624   gtex_v8
#> 38:                        Esophagus - Mucosa 2.05779e-04  0.471568   gtex_v8
#> 39:                        Esophagus - Mucosa 2.48068e-09  0.655426   gtex_v8
#> 40:                    Esophagus - Muscularis 3.54545e-10  0.984300   gtex_v8
#> 41:                    Esophagus - Muscularis 6.11870e-07 -0.721811   gtex_v8
#> 42:                    Esophagus - Muscularis 2.50817e-05  0.408854   gtex_v8
#> 43:                  Heart - Atrial Appendage 2.12053e-08  1.000880   gtex_v8
#> 44:                  Heart - Atrial Appendage 4.72450e-10 -1.089990   gtex_v8
#> 45:                  Heart - Atrial Appendage 1.34356e-10  0.905711   gtex_v8
#> 46:                    Heart - Left Ventricle 9.72458e-12 -1.034280   gtex_v8
#> 47:                    Heart - Left Ventricle 3.44863e-09  0.759825   gtex_v8
#> 48:                                     Liver 1.22632e-06 -0.992022   gtex_v8
#> 49:                                      Lung 1.32367e-07  0.747842   gtex_v8
#> 50:                                      Lung 4.68334e-05 -0.587194   gtex_v8
#> 51:                                      Lung 4.57721e-11  0.701631   gtex_v8
#> 52:                         Muscle - Skeletal 9.91823e-09 -0.620506   gtex_v8
#> 53:                         Muscle - Skeletal 1.45337e-14  0.666372   gtex_v8
#> 54:                            Nerve - Tibial 2.97619e-15  1.050980   gtex_v8
#> 55:                            Nerve - Tibial 2.01162e-11 -0.861511   gtex_v8
#> 56:                            Nerve - Tibial 1.25308e-13  0.653294   gtex_v8
#> 57:                                  Pancreas 1.48904e-05  0.771603   gtex_v8
#> 58:                                 Pituitary 5.39767e-05  0.882921   gtex_v8
#> 59:                                 Pituitary 6.13186e-08  0.945009   gtex_v8
#> 60:                                  Prostate 1.52886e-05  1.005930   gtex_v8
#> 61:       Skin - Not Sun Exposed (Suprapubic) 6.03296e-11  0.911987   gtex_v8
#> 62:       Skin - Not Sun Exposed (Suprapubic) 7.22930e-08 -0.770752   gtex_v8
#> 63:       Skin - Not Sun Exposed (Suprapubic) 1.72014e-14  0.699855   gtex_v8
#> 64:            Skin - Sun Exposed (Lower leg) 8.00361e-11  0.857182   gtex_v8
#> 65:            Skin - Sun Exposed (Lower leg) 3.30591e-09 -0.746043   gtex_v8
#> 66:            Skin - Sun Exposed (Lower leg) 1.23187e-15  0.688242   gtex_v8
#> 67:                                    Spleen 3.15980e-06  0.781740   gtex_v8
#> 68:                                   Stomach 1.57231e-05  0.625518   gtex_v8
#> 69:                                   Thyroid 7.31007e-08  0.767395   gtex_v8
#> 70:                                   Thyroid 6.40558e-08 -0.795998   gtex_v8
#> 71:                                   Thyroid 6.91838e-11  0.605093   gtex_v8
#>                              tissueSiteDetail      pValue       nes datasetId
 xQTLdownload_eqtlSig("11_66328719_T_C_b37", datasetId="gtex_v7")
#> == Check the variant name:
#> Querying variant... | 2022-6月-01 16:34:57 
#> == Done.
#> == Start downloading significant eQTL associations:
#> == Done.
#> =================================
#> Totally 68 associatons were found for variant [11_66328719_T_C_b37] in 29 tissues in gtex_v7.
#>               variantId    snpId          gencodeId    geneSymbol
#>  1: 11_66328719_T_C_b37 rs618838  ENSG00000254986.3          DPP3
#>  2: 11_66328719_T_C_b37 rs618838 ENSG00000174483.15          BBS1
#>  3: 11_66328719_T_C_b37 rs618838  ENSG00000269990.1 CTD-3074O7.12
#>  4: 11_66328719_T_C_b37 rs618838  ENSG00000248746.1         ACTN3
#>  5: 11_66328719_T_C_b37 rs618838  ENSG00000248746.1         ACTN3
#>  6: 11_66328719_T_C_b37 rs618838  ENSG00000269990.1 CTD-3074O7.12
#>  7: 11_66328719_T_C_b37 rs618838  ENSG00000174165.3       ZDHHC24
#>  8: 11_66328719_T_C_b37 rs618838  ENSG00000248746.1         ACTN3
#>  9: 11_66328719_T_C_b37 rs618838  ENSG00000174080.6          CTSF
#> 10: 11_66328719_T_C_b37 rs618838  ENSG00000248746.1         ACTN3
#> 11: 11_66328719_T_C_b37 rs618838  ENSG00000174080.6          CTSF
#> 12: 11_66328719_T_C_b37 rs618838 ENSG00000173715.11      C11orf80
#> 13: 11_66328719_T_C_b37 rs618838  ENSG00000250105.1  CTD-3074O7.2
#> 14: 11_66328719_T_C_b37 rs618838  ENSG00000250105.1  CTD-3074O7.2
#> 15: 11_66328719_T_C_b37 rs618838  ENSG00000250105.1  CTD-3074O7.2
#> 16: 11_66328719_T_C_b37 rs618838  ENSG00000248746.1         ACTN3
#> 17: 11_66328719_T_C_b37 rs618838  ENSG00000173992.4           CCS
#> 18: 11_66328719_T_C_b37 rs618838  ENSG00000248746.1         ACTN3
#> 19: 11_66328719_T_C_b37 rs618838  ENSG00000174080.6          CTSF
#> 20: 11_66328719_T_C_b37 rs618838  ENSG00000248746.1         ACTN3
#> 21: 11_66328719_T_C_b37 rs618838  ENSG00000174080.6          CTSF
#> 22: 11_66328719_T_C_b37 rs618838  ENSG00000248746.1         ACTN3
#> 23: 11_66328719_T_C_b37 rs618838  ENSG00000248746.1         ACTN3
#> 24: 11_66328719_T_C_b37 rs618838  ENSG00000174080.6          CTSF
#> 25: 11_66328719_T_C_b37 rs618838  ENSG00000255517.2  CTD-3074O7.5
#> 26: 11_66328719_T_C_b37 rs618838  ENSG00000254986.3          DPP3
#> 27: 11_66328719_T_C_b37 rs618838  ENSG00000174080.6          CTSF
#> 28: 11_66328719_T_C_b37 rs618838  ENSG00000248746.1         ACTN3
#> 29: 11_66328719_T_C_b37 rs618838  ENSG00000174080.6          CTSF
#> 30: 11_66328719_T_C_b37 rs618838  ENSG00000248746.1         ACTN3
#> 31: 11_66328719_T_C_b37 rs618838  ENSG00000174080.6          CTSF
#> 32: 11_66328719_T_C_b37 rs618838  ENSG00000248746.1         ACTN3
#> 33: 11_66328719_T_C_b37 rs618838  ENSG00000174080.6          CTSF
#> 34: 11_66328719_T_C_b37 rs618838  ENSG00000173992.4           CCS
#> 35: 11_66328719_T_C_b37 rs618838 ENSG00000174483.15          BBS1
#> 36: 11_66328719_T_C_b37 rs618838  ENSG00000174165.3       ZDHHC24
#> 37: 11_66328719_T_C_b37 rs618838  ENSG00000248746.1         ACTN3
#> 38: 11_66328719_T_C_b37 rs618838  ENSG00000174080.6          CTSF
#> 39: 11_66328719_T_C_b37 rs618838  ENSG00000255517.2  CTD-3074O7.5
#> 40: 11_66328719_T_C_b37 rs618838  ENSG00000269990.1 CTD-3074O7.12
#> 41: 11_66328719_T_C_b37 rs618838  ENSG00000248746.1         ACTN3
#> 42: 11_66328719_T_C_b37 rs618838  ENSG00000174080.6          CTSF
#> 43: 11_66328719_T_C_b37 rs618838  ENSG00000254986.3          DPP3
#> 44: 11_66328719_T_C_b37 rs618838  ENSG00000174080.6          CTSF
#> 45: 11_66328719_T_C_b37 rs618838 ENSG00000174483.15          BBS1
#> 46: 11_66328719_T_C_b37 rs618838  ENSG00000269990.1 CTD-3074O7.12
#> 47: 11_66328719_T_C_b37 rs618838  ENSG00000248746.1         ACTN3
#> 48: 11_66328719_T_C_b37 rs618838  ENSG00000173621.8         LRFN4
#> 49: 11_66328719_T_C_b37 rs618838 ENSG00000174483.15          BBS1
#> 50: 11_66328719_T_C_b37 rs618838  ENSG00000269990.1 CTD-3074O7.12
#> 51: 11_66328719_T_C_b37 rs618838  ENSG00000248746.1         ACTN3
#> 52: 11_66328719_T_C_b37 rs618838  ENSG00000174080.6          CTSF
#> 53: 11_66328719_T_C_b37 rs618838  ENSG00000173992.4           CCS
#> 54: 11_66328719_T_C_b37 rs618838  ENSG00000255468.2 RP11-867G23.8
#> 55: 11_66328719_T_C_b37 rs618838  ENSG00000255517.2  CTD-3074O7.5
#> 56: 11_66328719_T_C_b37 rs618838  ENSG00000248746.1         ACTN3
#> 57: 11_66328719_T_C_b37 rs618838  ENSG00000250105.1  CTD-3074O7.2
#> 58: 11_66328719_T_C_b37 rs618838  ENSG00000255468.2 RP11-867G23.8
#> 59: 11_66328719_T_C_b37 rs618838  ENSG00000254986.3          DPP3
#> 60: 11_66328719_T_C_b37 rs618838  ENSG00000269990.1 CTD-3074O7.12
#> 61: 11_66328719_T_C_b37 rs618838  ENSG00000248746.1         ACTN3
#> 62: 11_66328719_T_C_b37 rs618838  ENSG00000174080.6          CTSF
#> 63: 11_66328719_T_C_b37 rs618838  ENSG00000173621.8         LRFN4
#> 64: 11_66328719_T_C_b37 rs618838  ENSG00000254986.3          DPP3
#> 65: 11_66328719_T_C_b37 rs618838  ENSG00000174080.6          CTSF
#> 66: 11_66328719_T_C_b37 rs618838  ENSG00000173992.4           CCS
#> 67: 11_66328719_T_C_b37 rs618838  ENSG00000173653.3          RCE1
#> 68: 11_66328719_T_C_b37 rs618838  ENSG00000173621.8         LRFN4
#>               variantId    snpId          gencodeId    geneSymbol
#>                          tissueSiteDetail      pValue       nes datasetId
#>  1:                Adipose - Subcutaneous 1.07519e-04 -0.109360   gtex_v7
#>  2:                Adipose - Subcutaneous 9.87971e-06 -0.127072   gtex_v7
#>  3:                Adipose - Subcutaneous 8.87171e-06 -0.149663   gtex_v7
#>  4:                Adipose - Subcutaneous 2.29790e-15  0.510854   gtex_v7
#>  5:          Adipose - Visceral (Omentum) 7.93813e-12  0.485882   gtex_v7
#>  6:                         Adrenal Gland 1.37979e-05 -0.326874   gtex_v7
#>  7:                        Artery - Aorta 1.31606e-05  0.219139   gtex_v7
#>  8:                        Artery - Aorta 1.62827e-17  0.678968   gtex_v7
#>  9:                        Artery - Aorta 2.77687e-09  0.250382   gtex_v7
#> 10:                       Artery - Tibial 1.01772e-11  0.403614   gtex_v7
#> 11:                       Artery - Tibial 6.35543e-13  0.205183   gtex_v7
#> 12:                       Artery - Tibial 3.55217e-05  0.180017   gtex_v7
#> 13:       Brain - Caudate (basal ganglia) 1.17324e-09 -0.635378   gtex_v7
#> 14:         Brain - Cerebellar Hemisphere 2.09367e-11 -0.736031   gtex_v7
#> 15:                    Brain - Cerebellum 2.65579e-11 -0.722031   gtex_v7
#> 16:               Breast - Mammary Tissue 3.80392e-11  0.567130   gtex_v7
#> 17:   Cells - EBV-transformed lymphocytes 6.99751e-09 -0.605784   gtex_v7
#> 18:       Cells - Transformed fibroblasts 3.22861e-07  0.344885   gtex_v7
#> 19:       Cells - Transformed fibroblasts 1.15557e-11  0.159478   gtex_v7
#> 20:                       Colon - Sigmoid 1.88475e-12  0.649715   gtex_v7
#> 21:                       Colon - Sigmoid 4.96950e-07  0.225425   gtex_v7
#> 22:                    Colon - Transverse 3.14061e-07  0.361959   gtex_v7
#> 23: Esophagus - Gastroesophageal Junction 4.15574e-15  0.604421   gtex_v7
#> 24: Esophagus - Gastroesophageal Junction 9.84022e-07  0.298715   gtex_v7
#> 25:                    Esophagus - Mucosa 1.68521e-07 -0.361858   gtex_v7
#> 26:                    Esophagus - Mucosa 7.60741e-11 -0.167233   gtex_v7
#> 27:                    Esophagus - Mucosa 1.24794e-11  0.191076   gtex_v7
#> 28:                Esophagus - Muscularis 1.22176e-28  0.665133   gtex_v7
#> 29:                Esophagus - Muscularis 2.25016e-15  0.323908   gtex_v7
#> 30:              Heart - Atrial Appendage 3.01964e-08  0.462257   gtex_v7
#> 31:                Heart - Left Ventricle 1.36089e-06  0.230397   gtex_v7
#> 32:                                  Lung 2.74250e-05  0.285196   gtex_v7
#> 33:                                  Lung 1.49331e-08  0.168562   gtex_v7
#> 34:                                  Lung 1.31175e-05 -0.111668   gtex_v7
#> 35:                     Muscle - Skeletal 2.02544e-06 -0.163666   gtex_v7
#> 36:                     Muscle - Skeletal 1.12857e-05  0.149825   gtex_v7
#> 37:                     Muscle - Skeletal 1.55539e-95  0.745788   gtex_v7
#> 38:                     Muscle - Skeletal 3.62770e-06  0.126911   gtex_v7
#> 39:                        Nerve - Tibial 1.03880e-04 -0.243841   gtex_v7
#> 40:                        Nerve - Tibial 1.48897e-07 -0.218771   gtex_v7
#> 41:                        Nerve - Tibial 1.03891e-24  0.631500   gtex_v7
#> 42:                        Nerve - Tibial 1.46503e-10  0.202532   gtex_v7
#> 43:                              Pancreas 2.44983e-08 -0.237880   gtex_v7
#> 44:                             Pituitary 8.16036e-12  0.425542   gtex_v7
#> 45:   Skin - Not Sun Exposed (Suprapubic) 3.13139e-05 -0.132643   gtex_v7
#> 46:   Skin - Not Sun Exposed (Suprapubic) 4.04293e-05 -0.158047   gtex_v7
#> 47:   Skin - Not Sun Exposed (Suprapubic) 7.48831e-07  0.379148   gtex_v7
#> 48:   Skin - Not Sun Exposed (Suprapubic) 6.14818e-05 -0.181240   gtex_v7
#> 49:        Skin - Sun Exposed (Lower leg) 2.27572e-08 -0.210556   gtex_v7
#> 50:        Skin - Sun Exposed (Lower leg) 2.52339e-06 -0.169583   gtex_v7
#> 51:        Skin - Sun Exposed (Lower leg) 1.54629e-05  0.275059   gtex_v7
#> 52:        Skin - Sun Exposed (Lower leg) 7.13137e-05  0.115617   gtex_v7
#> 53:                                Spleen 1.35070e-08 -0.334750   gtex_v7
#> 54:                                Testis 6.13131e-05  0.257024   gtex_v7
#> 55:                                Testis 1.40201e-05  0.229280   gtex_v7
#> 56:                                Testis 1.77863e-34  0.643565   gtex_v7
#> 57:                                Testis 4.53674e-35 -0.685325   gtex_v7
#> 58:                               Thyroid 9.89876e-05  0.210391   gtex_v7
#> 59:                               Thyroid 6.33157e-08 -0.135224   gtex_v7
#> 60:                               Thyroid 1.28011e-08 -0.250984   gtex_v7
#> 61:                               Thyroid 2.86402e-24  0.607988   gtex_v7
#> 62:                               Thyroid 3.43313e-07  0.139455   gtex_v7
#> 63:                               Thyroid 4.28855e-05 -0.154567   gtex_v7
#> 64:                           Whole Blood 3.33291e-12 -0.189478   gtex_v7
#> 65:                           Whole Blood 6.94377e-14  0.245937   gtex_v7
#> 66:                           Whole Blood 1.99942e-08 -0.153217   gtex_v7
#> 67:                           Whole Blood 3.45968e-05 -0.123848   gtex_v7
#> 68:                           Whole Blood 1.43758e-05 -0.196104   gtex_v7
#>                          tissueSiteDetail      pValue       nes datasetId
 xQTLdownload_eqtlSig("11_66328719_T_C_b37", datasetId="gtex_v7",
                       tissueSiteDetail="Skin - Sun Exposed (Lower leg)")
#> == Check the variant name:
#> Querying variant... | 2022-6月-01 16:34:59 
#> == Done.
#> == Start downloading significant eQTL associations:
#> == Done.
#> =================================
#> Totally 4 associatons were found for variant [11_66328719_T_C_b37] in Skin - Sun Exposed (Lower leg) in gtex_v7.
#>              variantId    snpId          gencodeId    geneSymbol
#> 1: 11_66328719_T_C_b37 rs618838 ENSG00000174483.15          BBS1
#> 2: 11_66328719_T_C_b37 rs618838  ENSG00000269990.1 CTD-3074O7.12
#> 3: 11_66328719_T_C_b37 rs618838  ENSG00000248746.1         ACTN3
#> 4: 11_66328719_T_C_b37 rs618838  ENSG00000174080.6          CTSF
#>                  tissueSiteDetail      pValue       nes datasetId
#> 1: Skin - Sun Exposed (Lower leg) 2.27572e-08 -0.210556   gtex_v7
#> 2: Skin - Sun Exposed (Lower leg) 2.52339e-06 -0.169583   gtex_v7
#> 3: Skin - Sun Exposed (Lower leg) 1.54629e-05  0.275059   gtex_v7
#> 4: Skin - Sun Exposed (Lower leg) 7.13137e-05  0.115617   gtex_v7

 # Download eQTL association according to all tissues with genome location:
 varInfo <-  xQTLquery_varPos(chrom="chr1", pos=c(1102708),"gtex_v8")
#> Querying variant by position... | 2022-6月-01 16:35:00 
#>    Got varints: 1/1
 xQTLdownload_eqtlSig(variantName=varInfo$snpId)
#> == Check the variant name:
#> Querying variant... | 2022-6月-01 16:35:00 
#> == Done.
#> == Start downloading significant eQTL associations:
#> == Done.
#> =================================
#> Totally 20 associatons were found for variant [rs66622470] in 20 tissues in gtex_v8.
#>                variantId      snpId          gencodeId    geneSymbol
#>  1: chr1_1102708_G_C_b38 rs66622470 ENSG00000131591.17      C1orf159
#>  2: chr1_1102708_G_C_b38 rs66622470 ENSG00000131591.17      C1orf159
#>  3: chr1_1102708_G_C_b38 rs66622470 ENSG00000131591.17      C1orf159
#>  4: chr1_1102708_G_C_b38 rs66622470 ENSG00000131591.17      C1orf159
#>  5: chr1_1102708_G_C_b38 rs66622470 ENSG00000131591.17      C1orf159
#>  6: chr1_1102708_G_C_b38 rs66622470 ENSG00000131591.17      C1orf159
#>  7: chr1_1102708_G_C_b38 rs66622470 ENSG00000131591.17      C1orf159
#>  8: chr1_1102708_G_C_b38 rs66622470 ENSG00000131591.17      C1orf159
#>  9: chr1_1102708_G_C_b38 rs66622470 ENSG00000131591.17      C1orf159
#> 10: chr1_1102708_G_C_b38 rs66622470 ENSG00000131591.17      C1orf159
#> 11: chr1_1102708_G_C_b38 rs66622470 ENSG00000131591.17      C1orf159
#> 12: chr1_1102708_G_C_b38 rs66622470 ENSG00000131591.17      C1orf159
#> 13: chr1_1102708_G_C_b38 rs66622470 ENSG00000131591.17      C1orf159
#> 14: chr1_1102708_G_C_b38 rs66622470  ENSG00000223823.1     LINC01342
#> 15: chr1_1102708_G_C_b38 rs66622470 ENSG00000131591.17      C1orf159
#> 16: chr1_1102708_G_C_b38 rs66622470 ENSG00000131591.17      C1orf159
#> 17: chr1_1102708_G_C_b38 rs66622470 ENSG00000131591.17      C1orf159
#> 18: chr1_1102708_G_C_b38 rs66622470 ENSG00000131591.17      C1orf159
#> 19: chr1_1102708_G_C_b38 rs66622470 ENSG00000131591.17      C1orf159
#> 20: chr1_1102708_G_C_b38 rs66622470  ENSG00000217801.9 RP11-465B22.3
#>                              tissueSiteDetail      pValue       nes datasetId
#>  1:                          Brain - Amygdala 1.60336e-06 -0.459798   gtex_v8
#>  2:  Brain - Anterior cingulate cortex (BA24) 8.66605e-07 -0.454671   gtex_v8
#>  3:           Brain - Caudate (basal ganglia) 8.31578e-13 -0.472193   gtex_v8
#>  4:             Brain - Cerebellar Hemisphere 2.37523e-06 -0.329395   gtex_v8
#>  5:                        Brain - Cerebellum 7.84906e-08 -0.392030   gtex_v8
#>  6:                            Brain - Cortex 9.13735e-13 -0.516023   gtex_v8
#>  7:              Brain - Frontal Cortex (BA9) 1.20942e-10 -0.454262   gtex_v8
#>  8:                       Brain - Hippocampus 1.94890e-08 -0.440415   gtex_v8
#>  9:                      Brain - Hypothalamus 1.87997e-08 -0.444856   gtex_v8
#> 10: Brain - Nucleus accumbens (basal ganglia) 9.14774e-15 -0.463095   gtex_v8
#> 11:           Brain - Putamen (basal ganglia) 1.18332e-11 -0.530753   gtex_v8
#> 12:                        Esophagus - Mucosa 1.30624e-07 -0.149683   gtex_v8
#> 13:                    Esophagus - Muscularis 6.41193e-07 -0.145362   gtex_v8
#> 14:                                      Lung 6.87033e-05  0.264793   gtex_v8
#> 15:                            Nerve - Tibial 5.37315e-05 -0.125807   gtex_v8
#> 16:                                 Pituitary 1.68798e-05 -0.294973   gtex_v8
#> 17:       Skin - Not Sun Exposed (Suprapubic) 4.95643e-10 -0.189245   gtex_v8
#> 18:            Skin - Sun Exposed (Lower leg) 1.67550e-07 -0.148698   gtex_v8
#> 19:                                   Thyroid 1.47633e-09 -0.183078   gtex_v8
#> 20:                               Whole Blood 4.51565e-05 -0.218352   gtex_v8

 # Download eQTL info for gene:
 xQTLdownload_eqtlSig(genes="ATAD3B")
#> == Check the gene name:
#> Querying genes... | 2022-6月-01 16:35:01 
#> 1 record has been obtained!
#> == Done.
#> == Start downloading significant eQTL associations:
#> == Done.
#> =================================
#> Totally 9175 associatons were found for gene: [ATAD3B] in 49 tissues in gtex_v8.
#>                   variantId       snpId          gencodeId geneSymbol
#>    1:  chr1_1177351_G_A_b38 rs186127282 ENSG00000160072.19     ATAD3B
#>    2:  chr1_1184747_C_T_b38  rs57934946 ENSG00000160072.19     ATAD3B
#>    3:  chr1_1185441_C_T_b38   rs3813205 ENSG00000160072.19     ATAD3B
#>    4:  chr1_1189594_C_T_b38  rs78717314 ENSG00000160072.19     ATAD3B
#>    5:  chr1_1192181_T_C_b38 rs115516254 ENSG00000160072.19     ATAD3B
#>   ---                                                                
#> 9171:  chr1_1580310_T_A_b38 rs191138129 ENSG00000160072.19     ATAD3B
#> 9172:  chr1_1581288_T_C_b38 rs116579874 ENSG00000160072.19     ATAD3B
#> 9173: chr1_1582313_A_AC_b38 rs146182505 ENSG00000160072.19     ATAD3B
#> 9174:  chr1_1582336_A_G_b38  rs77672677 ENSG00000160072.19     ATAD3B
#> 9175:  chr1_2024222_C_T_b38 rs191110510 ENSG00000160072.19     ATAD3B
#>             tissueSiteDetail      pValue       nes datasetId
#>    1: Adipose - Subcutaneous 4.49057e-14 -1.114530   gtex_v8
#>    2: Adipose - Subcutaneous 4.12312e-05 -0.274952   gtex_v8
#>    3: Adipose - Subcutaneous 2.27961e-09 -0.742035   gtex_v8
#>    4: Adipose - Subcutaneous 2.27961e-09 -0.742035   gtex_v8
#>    5: Adipose - Subcutaneous 2.77360e-12 -0.822050   gtex_v8
#>   ---                                                       
#> 9171:            Whole Blood 2.24586e-14 -0.565012   gtex_v8
#> 9172:            Whole Blood 1.17326e-14 -0.557188   gtex_v8
#> 9173:            Whole Blood 2.90627e-16 -0.670756   gtex_v8
#> 9174:            Whole Blood 2.90627e-16 -0.670756   gtex_v8
#> 9175:            Whole Blood 1.02499e-04  0.321080   gtex_v8
 xQTLdownload_eqtlSig(genes=c("TP53", "SLC35E2B"), tissueSiteDetail= "Brain - Cerebellum")
#> == Check the gene name:
#> Querying genes... | 2022-6月-01 16:35:05 
#> Batch 1/1. Downloaded  page1/1; 2 records.
#> == Done.
#> == Start downloading significant eQTL associations:
#> == Done.
#> =================================
#> Totally 127 associatons were found for gene: [TP53, SLC35E2B] in Brain - Cerebellum in gtex_v8.
#>                     variantId       snpId          gencodeId geneSymbol
#>   1:     chr1_1584267_G_A_b38 rs145042579 ENSG00000189339.11   SLC35E2B
#>   2:     chr1_1584321_T_C_b38   rs9442406 ENSG00000189339.11   SLC35E2B
#>   3:     chr1_1585345_C_T_b38  rs28788874 ENSG00000189339.11   SLC35E2B
#>   4:     chr1_1586425_A_G_b38   rs4648786 ENSG00000189339.11   SLC35E2B
#>   5:     chr1_1586626_C_T_b38   rs4648787 ENSG00000189339.11   SLC35E2B
#>  ---                                                                   
#> 123:     chr1_1674868_G_A_b38  rs61776786 ENSG00000189339.11   SLC35E2B
#> 124:     chr1_1676790_C_T_b38  rs61776790 ENSG00000189339.11   SLC35E2B
#> 125:     chr1_1677720_T_G_b38  rs12724172 ENSG00000189339.11   SLC35E2B
#> 126:     chr1_1677726_A_T_b38  rs12757736 ENSG00000189339.11   SLC35E2B
#> 127: chr1_1730067_TGACA_T_b38 rs200238609 ENSG00000189339.11   SLC35E2B
#>        tissueSiteDetail      pValue       nes datasetId
#>   1: Brain - Cerebellum 1.56677e-05 -1.202540   gtex_v8
#>   2: Brain - Cerebellum 1.95486e-05 -0.457280   gtex_v8
#>   3: Brain - Cerebellum 6.50674e-15 -0.732374   gtex_v8
#>   4: Brain - Cerebellum 8.59133e-12 -0.625619   gtex_v8
#>   5: Brain - Cerebellum 8.59133e-12 -0.625619   gtex_v8
#>  ---                                                   
#> 123: Brain - Cerebellum 2.74194e-10  0.560539   gtex_v8
#> 124: Brain - Cerebellum 3.26025e-10  0.550227   gtex_v8
#> 125: Brain - Cerebellum 3.29377e-10  0.554540   gtex_v8
#> 126: Brain - Cerebellum 3.29377e-10  0.554540   gtex_v8
#> 127: Brain - Cerebellum 1.26892e-19 -1.282360   gtex_v8
 xQTLdownload_eqtlSig(genes="ENSG00000141510.16", datasetId="gtex_v8")
#> == Check the gene name:
#> Querying genes... | 2022-6月-01 16:35:06 
#> 1 record has been obtained!
#> == Done.
#> == Start downloading significant eQTL associations:
#> == Done.
#> =================================
#> Totally 33 associatons were found for gene: [ENSG00000141510.16] in 7 tissues in gtex_v8.
#>                  variantId       snpId          gencodeId geneSymbol
#>  1:  chr17_7559785_G_A_b38 rs143094271 ENSG00000141510.16       TP53
#>  2:  chr17_7668434_T_G_b38  rs78378222 ENSG00000141510.16       TP53
#>  3:  chr17_7675353_C_T_b38  rs35850753 ENSG00000141510.16       TP53
#>  4:  chr17_7668434_T_G_b38  rs78378222 ENSG00000141510.16       TP53
#>  5:  chr17_7668434_T_G_b38  rs78378222 ENSG00000141510.16       TP53
#>  6:  chr17_7441178_C_T_b38   rs4151125 ENSG00000141510.16       TP53
#>  7:  chr17_7447656_T_C_b38   rs2302765 ENSG00000141510.16       TP53
#>  8:  chr17_7447918_G_A_b38   rs4796417 ENSG00000141510.16       TP53
#>  9:  chr17_7449438_T_G_b38 rs111647739 ENSG00000141510.16       TP53
#> 10:  chr17_7449553_A_G_b38 rs112623133 ENSG00000141510.16       TP53
#> 11:  chr17_7449917_G_A_b38 rs148251796 ENSG00000141510.16       TP53
#> 12:  chr17_7451596_C_T_b38  rs28456260 ENSG00000141510.16       TP53
#> 13:  chr17_7452340_G_A_b38  rs12602551 ENSG00000141510.16       TP53
#> 14:  chr17_7453283_A_G_b38  rs12938637 ENSG00000141510.16       TP53
#> 15:  chr17_7454681_C_G_b38   rs4796418 ENSG00000141510.16       TP53
#> 16:  chr17_7455958_T_C_b38   rs2302763 ENSG00000141510.16       TP53
#> 17:  chr17_7456438_G_A_b38   rs9901900 ENSG00000141510.16       TP53
#> 18:  chr17_7457004_T_C_b38   rs3855924 ENSG00000141510.16       TP53
#> 19:  chr17_7458740_C_T_b38 rs112051593 ENSG00000141510.16       TP53
#> 20:  chr17_7464046_G_A_b38   rs3867593 ENSG00000141510.16       TP53
#> 21: chr17_7464362_GT_G_b38  rs34739055 ENSG00000141510.16       TP53
#> 22:  chr17_7487350_A_G_b38  rs10432029 ENSG00000141510.16       TP53
#> 23:  chr17_7502618_C_T_b38   rs2071504 ENSG00000141510.16       TP53
#> 24:  chr17_7665310_G_A_b38 rs117557561 ENSG00000141510.16       TP53
#> 25:  chr17_7559785_G_A_b38 rs143094271 ENSG00000141510.16       TP53
#> 26:  chr17_7668434_T_G_b38  rs78378222 ENSG00000141510.16       TP53
#> 27:  chr17_7675353_C_T_b38  rs35850753 ENSG00000141510.16       TP53
#> 28:  chr17_7650924_T_C_b38   rs1641513 ENSG00000141510.16       TP53
#> 29:  chr17_7668434_T_G_b38  rs78378222 ENSG00000141510.16       TP53
#> 30:  chr17_7668783_C_T_b38  rs17884306 ENSG00000141510.16       TP53
#> 31:  chr17_7671403_C_T_b38  rs17879377 ENSG00000141510.16       TP53
#> 32:  chr17_7675353_C_T_b38  rs35850753 ENSG00000141510.16       TP53
#> 33:  chr17_7796745_C_T_b38 rs138420351 ENSG00000141510.16       TP53
#>                  variantId       snpId          gencodeId geneSymbol
#>                        tissueSiteDetail      pValue       nes datasetId
#>  1:              Adipose - Subcutaneous 1.00467e-05 -0.403825   gtex_v8
#>  2:              Adipose - Subcutaneous 8.68606e-10 -0.765786   gtex_v8
#>  3:              Adipose - Subcutaneous 1.34916e-05 -0.401123   gtex_v8
#>  4:        Cells - Cultured fibroblasts 9.07463e-06 -0.473760   gtex_v8
#>  5:                  Esophagus - Mucosa 5.18785e-07 -0.729226   gtex_v8
#>  6:                     Kidney - Cortex 7.27802e-07 -0.541302   gtex_v8
#>  7:                     Kidney - Cortex 7.27802e-07 -0.541302   gtex_v8
#>  8:                     Kidney - Cortex 7.27802e-07 -0.541302   gtex_v8
#>  9:                     Kidney - Cortex 8.46971e-07 -0.612803   gtex_v8
#> 10:                     Kidney - Cortex 7.27802e-07 -0.541302   gtex_v8
#> 11:                     Kidney - Cortex 7.27802e-07 -0.541302   gtex_v8
#> 12:                     Kidney - Cortex 6.52895e-07 -0.539388   gtex_v8
#> 13:                     Kidney - Cortex 7.27802e-07 -0.541302   gtex_v8
#> 14:                     Kidney - Cortex 7.27802e-07 -0.541302   gtex_v8
#> 15:                     Kidney - Cortex 7.27802e-07 -0.541302   gtex_v8
#> 16:                     Kidney - Cortex 7.27802e-07 -0.541302   gtex_v8
#> 17:                     Kidney - Cortex 7.27802e-07 -0.541302   gtex_v8
#> 18:                     Kidney - Cortex 7.27802e-07 -0.541302   gtex_v8
#> 19:                     Kidney - Cortex 7.27802e-07 -0.541302   gtex_v8
#> 20:                     Kidney - Cortex 4.64170e-06  0.514631   gtex_v8
#> 21:                     Kidney - Cortex 4.64170e-06  0.514631   gtex_v8
#> 22:                     Kidney - Cortex 4.69866e-06 -0.507172   gtex_v8
#> 23:                     Kidney - Cortex 3.96509e-06 -0.506233   gtex_v8
#> 24:                           Pituitary 5.70716e-07 -1.168270   gtex_v8
#> 25: Skin - Not Sun Exposed (Suprapubic) 2.88317e-05 -0.284258   gtex_v8
#> 26: Skin - Not Sun Exposed (Suprapubic) 2.05198e-10 -0.581091   gtex_v8
#> 27: Skin - Not Sun Exposed (Suprapubic) 8.94942e-06 -0.307209   gtex_v8
#> 28:      Skin - Sun Exposed (Lower leg) 4.51044e-05  0.101825   gtex_v8
#> 29:      Skin - Sun Exposed (Lower leg) 3.00426e-11 -0.562542   gtex_v8
#> 30:      Skin - Sun Exposed (Lower leg) 1.01029e-04 -0.148417   gtex_v8
#> 31:      Skin - Sun Exposed (Lower leg) 2.78866e-05 -0.172719   gtex_v8
#> 32:      Skin - Sun Exposed (Lower leg) 2.65464e-05 -0.255301   gtex_v8
#> 33:      Skin - Sun Exposed (Lower leg) 4.27947e-06 -0.406142   gtex_v8
#>                        tissueSiteDetail      pValue       nes datasetId
 xQTLdownload_eqtlSig(genes="ENSG00000141510.11", datasetId="gtex_v7",
                      tissueSiteDetail="Thyroid" )
#> == Check the gene name:
#> Querying genes... | 2022-6月-01 16:35:08 
#> 1 record has been obtained!
#> == Done.
#> == Start downloading significant eQTL associations:
#> == Done.
#> =================================
#> Totally 2 associatons were found for gene: [ENSG00000141510.11] in Thyroid in gtex_v7.
#>             variantId       snpId          gencodeId geneSymbol
#> 1: 17_7478402_C_G_b37  rs78414518 ENSG00000141510.11       TP53
#> 2: 17_7492388_G_A_b37 rs117372143 ENSG00000141510.11       TP53
#>    tissueSiteDetail      pValue       nes datasetId
#> 1:          Thyroid 1.88616e-05 -0.249789   gtex_v7
#> 2:          Thyroid 1.03285e-05 -0.255895   gtex_v7

 # Download eQTL info for a variant-gene pair:
 xQTLdownload_eqtlSig(variantName="rs1641513", genes="TP53", datasetId="gtex_v8")
#> == Check the variant name:
#> Querying variant... | 2022-6月-01 16:35:09 
#> == Done.
#> == Check the gene name:
#> Querying genes... | 2022-6月-01 16:35:09 
#> 1 record has been obtained!
#> == Done.
#> == Start downloading significant eQTL associations:
#> == Done.
#> =================================
#> Totally 1 associatons were found for variant [rs1641513] - gene: [TP53] in 1 tissue in gtex_v8.
#>                variantId     snpId          gencodeId geneSymbol
#> 1: chr17_7650924_T_C_b38 rs1641513 ENSG00000141510.16       TP53
#>                  tissueSiteDetail      pValue      nes datasetId
#> 1: Skin - Sun Exposed (Lower leg) 4.51044e-05 0.101825   gtex_v8
 xQTLdownload_eqtlSig(variantName="rs1641513", genes="TP53", datasetId="gtex_v7")
#> == Check the variant name:
#> Querying variant... | 2022-6月-01 16:35:10 
#> == Done.
#> == Check the gene name:
#> Querying genes... | 2022-6月-01 16:35:11 
#> 1 record has been obtained!
#> == Done.
#> == Start downloading significant eQTL associations:
#> No significant associations were found for variant [rs1641513]- gene [TP53] in 0 tissues in gtex_v7
#> Null data.table (0 rows and 0 cols)
 xQTLdownload_eqtlSig(variantName="chr1_1667948_A_G_b38",
                   genes="SLC35E2B", tissueSiteDetail="Kidney - Cortex")
#> == Check the variant name:
#> Querying variant... | 2022-6月-01 16:35:13 
#> == Done.
#> == Check the gene name:
#> Querying genes... | 2022-6月-01 16:35:14 
#> 1 record has been obtained!
#> == Done.
#> == Start downloading significant eQTL associations:
#> == Done.
#> =================================
#> Totally 1 associatons were found for variant [chr1_1667948_A_G_b38] - gene: [SLC35E2B] in Kidney - Cortex in gtex_v8.
#>               variantId       snpId          gencodeId geneSymbol
#> 1: chr1_1667948_A_G_b38 rs368438548 ENSG00000189339.11   SLC35E2B
#>    tissueSiteDetail      pValue      nes datasetId
#> 1:  Kidney - Cortex 7.21369e-14 -1.05597   gtex_v8
# }