xQTLquery_gene.

users can use this function to query basic information (including gene name, symbol, position and description, etc ) of genes.

Usage

xQTLquery_gene(
  genes = "",
  geneType = "auto",
  gencodeVersion = "v26",
  recordPerChunk = 150
)

Arguments

genes

A charater vector or a string of gene symbol, gencode id (versioned or unversioned), or a charater string of gene type.

gene symbol (Default).

A character string or a character vector (case ignored). like: "tp53","naDK","SDF4".

gencode/ensemble id (versioned or unversioned).

A character string or a character vector (case ignored). like: "ENSG00000210195.2","ENSG00000078808".

gene classification.

when "geneType" is "geneCategory", supported "genes" can be listed using function gencodeGenetype$V26 or gencodeGenetype$V19

geneType

A character string. "auto"(default), "geneSymbol", "gencodeId" or "geneCategory".

gencodeVersion

"v26" or "v19"

recordPerChunk

A integer value. Defaulut: 150

Value

A data.table of queried gene information. including following columns:

genes. Input genes

geneSymbol. Gene symbol.

gencodeId. Gencode/ensemble id (versioned).

entrezGeneId. Entrez gene ID.

geneType. Gene type.

chromosome. Note: "chr" is added in gencode v26,

start.

end.

strand.

tss. Transcriptional start site.

gencodeVersion. Gencode Version.

genomeBuild. Genome version.

description.

Examples

# \donttest{
 # get all protein coding genes and description:
 # protein_coding <- xQTLquery_gene(genes="protein coding")
 # get all miRNA and description:
 # miRNA <- xQTLquery_gene(genes="miRNA", geneType="geneCategory", "v19")

 # hg38 test:
 geneInfo <- xQTLquery_gene("TP53")
#> Querying genes... | 2022-6月-01 16:36:19 
#> 1 record has been obtained!
 geneInfo <- xQTLquery_gene(c("tp53","naDK","SDF4") )
#> Querying genes... | 2022-6月-01 16:36:20 
#> Batch 1/1. Downloaded  page1/1; 3 records.
 geneInfo <- xQTLquery_gene(c("ENSG00000210195.2","ENSG00000078808"))
#> Querying genes... | 2022-6月-01 16:36:20 
#> Batch 1/1. Downloaded  page1/1; 2 records.

 # hg19 test:
 geneInfo <- xQTLquery_gene(c("TP53","naDK"),  gencodeVersion="v19")
#> Querying genes... | 2022-6月-01 16:36:21 
#> Batch 1/1. Downloaded  page1/1; 2 records.
 geneInfo <- xQTLquery_gene(c("ENSG00000141510.11","ENSG00000008130.11"), gencodeVersion="v19")
#> Querying genes... | 2022-6月-01 16:36:22 
#> Batch 1/1. Downloaded  page1/1; 2 records.

 # }